// src/types/post.ts
// Generated by https://quicktype.io

import type { University } from "./university";
import type { User } from "./user";

export interface Post {
  "post-id": number;
  "user-id": number;
  date: string;
  "post-description": string;
  isbn: string;
  "uni-id": number;
  price: string;
  status: number;
  boosts: any[];
  university: University;
  user: User;
}